+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+- Upsert flags when loading feature flags with computed errors ([#38](https://github.com/PostHog/posthog-android/pull/38)) 
+
 ## 3.0.0-alpha.5 - 2023-10-04
 
 - Does not report screen events if there's no title ([#32](https://github.com/PostHog/posthog-android/pull/32))